add tunable to read block via partial object (#184)
Add a tunable which enables reading blocks from S3 by getting first the
DataObject header, and then just the block we are looking for.  Disabled
by default but shows that this works.

In the future we can figure out when it would be a net performance
benefit to do this.

Gets around 5x more random blocks per second, while consuming 1/10th the
network bandwidth and 2/3 the CPU  (requires S3 queue depth=300).