write_block without spawning a new task (#296)
We can process the write_block request without spawning a new task.
Instead we execute a (non-async) callback when the block has been
persisted to S3.  This improves performance due to reducing load on the
tokio task scheduler, and reducing contention on the pool's
syncing_state lock.  It runs `Pool::write_block()` from the
`Server::start_connection()` task, but since the the CPU time is
minimal, and the CPU portion of `Pool::write_block()` is executed
serially anyway (due to needing the syncing_state lock).