use rwlock for pool guid mapping (#305)
For high-IOPS workloads, the ZettaCache state lock is highly contended.
We need to acquire it for every lookup and insert.  Because the
PoolGuid->PoolId mapping is inside the ZettaCacheState, we need to
acquire the lock multiple times for each lookup().  The PoolGuid ->
PoolId mapping is rarely changed.

This commit changes the PoolGuid -> PoolId mapping to be protected by
its own RwLock, instead of being inside the ZettaCacheState.  Now
lookup() only needs to acquire the ZettaCacheState lock once.

Performance of sequential reads increases by 10-15%.

There's no on-disk format change, because the PoolGuidMappingPhys is
represented transparently because it's a newtype.

Bonus change: The ZettaCache is cloned for every block that's inserted
via insert_all(), which clones each of its many members.  This commit
puts the whole ZettaCache in its own Arc, so that it can be cloned with
a single atomic op.