Measurement.fut() doubles future size (#306)
`Measurement.fut()` takes a future and returns a future which wraps it.
However, doing this with an async closure ends up doubling the size of
the future (i.e. the returned future is double the size of the passed-in
future).  If there are several layers of this, it can result in very
large futures, which are expensive to create and copy around.

This commit resolves the problem by using a simple future combinator,
`FutureExt::inspect()`, to wrap the provided future.  Now the returned
future is just 8 bytes larger than the provided one.

A test is added to ensure that `.fut()` does not increase the future
size exponentially.