DLPX-80232 zettacache healing due to insert just before rebalance (#321)
Immediately after each checkpoint completes, we call
`try_start_merge_task()` which may kick off background tasks to do an
index merge.  Those tasks include copying any blocks that need to be
moved as part of rebalancing/evacuation.

* Code path `checkpoint_task()` -> `try_start_merge_task()` ->
  `spawn_merge_task()` -> `merge_task()` -> `rebalance()`

The "kicking off" happens with the ZettaCacheState lock held, but we may
acquire the lock at any arbitrary point.  In particular, there may be
insertions that are not yet reflected in the on-disk checkpoint, and the
disk writes for those insertions may not have completed yet.

* `ZettaCacheState::insert()` has returned, but the future it returns
  (which calls `BlockAccess::write_raw()`) has not yet completed.

According to the ZettaCacheState, this newly-inserted block is part of
the cache (e.g. it is allocated according to the BlockAllocator).  If
rebalance decides to evacuate the slab that it's in, we need to allocate
a new place for this block, and copy its contents to the new location.

However, because the insert's write may not have completed,
`MergeState::rebalance()` may read the wrong (old) data from disk.  A
later lookup of this block will return incorrect data to the kernel.
The kernel detects that the checksum is incorrect, and retries the
read_block with the `heal` flag set, which causes the agent to get the
object from S3 and compare its contents to what's in the zettacache.
The agent notices that the contents differ, so it "heals" the zettacache
and returns the S3 data to the kernel, which has the correct checksum.

The problem can be more easily reproduced by adding delays before
`try_start_merge_task()`, and `ZettaCacheState::insert()`'s call to
`BlockAccess::write_raw()`.

This PR fixes the problem by having
`ZettaCacheState::try_start_merge_task()` wait for all outstanding
writes to disk, if we are starting a rebalance.

Note that after try_start_merge_task() drops the ZettaCacheState lock,
new inserts can not allocate from evacuating slabs, and therefore can
not be part of this rebalance.

Bonus change: reorganize and better log the reason for starting a merge.