periodically update track_instantaneous() stats (#320)
Some stats need to be updated manually with track_instantaneous(),
specifically:

  BlockAllocatorSize
  BlockAllocatorAvailable
  BlockAllocatorFreeSlabsSize
  PendingChanges
  SpeculativeBufferBytesAvailable
  DemandBufferBytesAvailable

There's a different approach to updating each of these, leading to a few
problems:

BlockAllocatorAvailable/FreeSlabsSize is updated only when starting a
checkpoint, so it's value is out of date between checkpoints.
Additionally, it may not reflect a "typical" value because the
measurement is taken just before these stats change due to the
checkpoint.  E.g. If the free slabs are exhausted each checkpoint,
FreeSlabsSize may always show as 0, even though it's going up and down.

Speculative/DemandBufferBytesAvailable is not updated when there's no
activity, so it misses the last update before activity ceases. (see
DLPX-80032)

This commit consolidates the updating of all these stats to a single,
periodic task, which runs once a second by default.