zettacache doesn't need to lookup before inserting writes (#328)
When inserting newly-written blocks to the zettacache, there's a slight
chance that the BlockId may already be in the cache, due to a previous
kernel crash.  In this case, instead of doing a lookup and omitting the
new block from the cache, we can insert over top of the old version of
this block.

This saves us 8% of one CPU, during write ingestion, and pressure on the
ZettaCacheState lock.