scrub reads should not alter LRU order (#331)
A `zpool scrub` will read every block in the pool.  This will update the
Atime of all blocks in the zettacache, changing their LRU order.
However, this doesn't reflect the "typical" workload.

This commit makes scrub reads (those with the `heal` flag set) not
update the blocks' Atimes, or record their access in the hits-by-size
histogram.  They are still visible in `zcache stats`.

Bonus change: extensive code cleanup in the area.