zettacache may read/write past end of disk (#343)
When the zettacache is created, the initial Extent of the entire disk
capacity is miscalculated to be 4KB too long.  Typically this has no ill
effect because there's padding at the end of the disk to align it to a
multiple of the slab size (16MB).  But if the disk size is exactly
right, there may be no slab padding, in which case we'll try to
read/write 4KB past the end of the disk, resulting in an assertion
failure.

This commit corrects the Extent calculation.  Note that the calculation
performed when adding a disk was already correct.