Redesign On-Disk Slab Type Structures (#334)
== Motivation

Every checkpoint in the ZettaCache we write a ZettaCheckpointPhys to
disk by generating a bunch of in-memory state in JSON and then
compressing that JSON blob with LZ4 before saving it on-disk. The bigger
this structure is the more time is spent on converting it and storing it
to disk. The biggest portion of this structure consists of a vector
containing a SlabPhys for each slab in our block allocator that
describes the slab’s type (this is used to reconstruct the slab’s state
when importing a zettacache).

As the size of the ZettaCache grows and the number of slabs increases,
so does the size of the checkpoints that we write every
minute/checkpoint. For caches that we use internally [24~256GB] this is
not as big of a problem, but for customer systems (current expectation
in the order of 10~100TB) we expect to hit problems. Not only we’ll be
wasting a lot of time converting our checkpoint structures to JSON and
compressing them, but we’ll be blocking almost every other operation in
the cache because we’re holding the cache’s state lock while doing so.

This patch proposes a design to minimize the aforementioned overheads.

= Current Design

Every checkpoint we recreate the ZettaCheckpointPhys structure before we
write it to disk. This structure includes a BlockAllocatorPhys that’s
created on demand during BlockAllocator::flush(). That function iterates
through all the slab structures in memory generating a SlabPhys for each
and pushing that into a newly created vector that will be assigned to
the `slabs` field of our BlockAllocatorPhys for the current checkpoint
being flushed.

The problem with the current design is that we write a vector with all
`SlabPhys` structures from all the slabs at every checkpoint to disk,
essentially recreating that state (regardless of how many slabs actually
changed their type).

To give an example, in a 100TB pool (filled up to 256GB) we have ~6.4M
slabs which generate a ~190MB which is later converted and compressed to
a ~2MB chunk.  This conversion takes close to 2~3s and it happens every
checkpoint. It also holds the zettacache state lock so no other
operation can take place at the same time.

Ideally we’d have a persistent on-disk structure that we update
incrementally with slab type changes that happened in the current
checkpoint.

= Proposed Design

Enhance the spacemap entries that bump a slab's generation to include
its type informat (e.g. `SlabPhys`) and get rid of the `slabs_phys`
field from the BlockAllocatorPhys struct. The idea is that every time a
slab changes its type or has its generation bumped, we log that change
in our space maps.

One disadvantage of this design is that during import we can't know the
generations of our slabs beforehand and thus can't skip obsolete entries
(e.g. everything is replayed). In our limited practice so far though we
haven't had this being a problem as the number of obsolete entries
hasn't been that overwhelming.

= Performance Of Proposed Design

Using stock bits In a 100TB sparse cache setup filling up a 256GB disk
we've found that that chunk_to_raw() would yield a compressed checkpoint
in ~890KB in 1~2 seconds. Before the cache run out of space (~250GB) the
same stock bits yielded ~1.4MB in 1.7~3.2 seconds. At that point
checkpoints would take 2.1~4.4 seconds.

With the bits of this proposed design on the same setup we started with
a compressed checkpoint of 1K in 0ms and when the cache run out of space
with 10K in 0~3ms. At that point checkpoints would take 301~700ms.

In terms of import time, on the above setup both VMs had almost
identical numbers with slab array initialization taking ~500ms and
reading ~1MB (~100K entries) of spacemap data taking 23ms.

= Alternatives Considered

Instead of appending these entries to our spacemaps use a dedicated
BlockBasedLog. The benefit of this approach over the proposed one's is
that we only replay the slab_type history during import and then knowing
the slab generations beforehand we can skip obsolete entries as before.
The main downside of this approach though is that we'd need to make sure
that the BBL doesn't grow indefinitely and gets condensed. We've already
put a lot of thought in spacemap condensing so the proposed approach is
already taking care of that for us. As a result we decided to not go
with this alternative.