DLPX-80638 zettacache healing due to outstanding_writes (#356)
When we initiate a checkpoint, in `ZettaCache::flush_checkpoint()`, we
need to wait for in-progress reads and writes, to ensure that they are
committed before the checkpoint is committed.

The problem is that after we wait for in-progress i/o, more i/o's can be
initiated before we acquire the zettacache state lock and take the
checkpoint.

By injecting delays and panic's, I was able to cause the agent to die
after the checkpoint was committed (i.e. the superblocks were written)
but before all the writes that should be part of it were completed.
When those blocks are subsequently read, a checksum error is detected
and the zettacache is healed.

This commit solves the problem by waiting for in-progress i/o
(`ZettaCacheState::outstanding_reads/_writes`) while holding the state
lock.  This increases the state lock hold time, although the lockless
wait still reduces the hold time somewhat.  This is not great for
performance, but it's a simple fix.