handle errors in Responder::response_task() (#355)
If the other side (e.g. the `zcache` command) closes the socket, we may
discover that and get an error when trying to read, or when trying to
write.  If Responder::response_task() gets an error when writing the
response, it will unwrap() and panic.

This commit changes Responder::response_task() to send the error back to
start_connection(), which will clean up the connection.