DLPX-80826 die_mtbf: new checkpoint may overwrite allocated slab (#376)
The slab allocator accumulates a list of "freeing" slabs each
checkpoint.  These are the slabs which are being freed in the current
checkpoint.  `SlabAllocator::checkpoint_done()` is called to indicate
that the checkpoint has been persisted and these slabe can be reused for
a different purpose.  However, they can not be reused before that,
because they are still in use according to the currently-persisted
checkpoint.

The problem is that `checkpoint_done()` is being called twice each
checkpoint, once before `PrimaryPhys::write_all()`.  After that call, we
allocate one more slab, for the checkpoint itself, via
`CheckpointPhys::write()`.  If that allocation happens to return one of
the `freeing` slabs that was incorrectly moved to `allocatable`, then
the checkpoint write will overwrite in-use data.  If the agent crashes
after that overwrite and before the PrimaryPhys is written, the
zettacache will be corrupted.  That said, the window here is very short,
and in practice we've only seen it when an agent crash is induced in
this narrow window via the `die_mtbf` testing tunable.  If the
corruption is of metadata (i.e. a slab of a BlockBasedLog) rather than
data (i.e. a slab from the BlockAllocator), then the corruption will be
non-recoverable, and the zettacache will need to be reset.

This commit removes the incorrect call to `checkpoint_done()`.  It was
added for its side effect of randomizing the allocatable slabs, but we
don't need to do that twice either.