code cleanup of BlockBasedLog (#373)
Various minor cleanup of BlockBasedLog code:

Annote entire struct with `#[serde(bound = "T: DeserializeOwned")]` to
indicate that the BlockBasedLogEntry needs to be deserializable without
capturing a reference to the input buffer (i.e. it is not
`Deserialize<'de>`).

Use `#[derivative(Default(bound = "T:"))]` to derive Default rather than
manually implementing it.

Create helper methods to iterate over the Extents that compose the
BlockBasedLog.

ReadOnlySummarizedBlockBasedLog: add a member for the entire Phys,
replacing the individual members copied from the phys.