fix performance of BlockBasedLogPhys::next_extent_to_write() (#383)
`BlockBasedLogPhys::next_extent_to_write()` uses `Iterator::last()` to
find the last extent.  Unfortunately, this is O(n), iterating over all
the elements to find the last one, even for `DoubleEndedIterator`s.
This can have a large impact on the performance of merging a large index
(where `n` is nontrivial).

This commit changes it to use `DoubleEndedIterator::next_back()`, which
is O(1).

The problem was introduced by #373.