fix divide-by-zero in BlockAllocator::flush() (#386)
If a merge completes when the zettacache is completely empty (no blocks
allocated), we can hit a divide-by-zero in BlockAllocator::flush() when
creating a log message.  This is difficult to hit in practice since we
wouldn't normally initiate a merge if there's nothing allocated.
However, future work will make it possible for this to happen.

This commit avoids the divide-by-zero by incrementing the divisor, which
is typically very large, so the impact on accuracy is negligible.