reduce memory usage of IndexRun by only storing IndexKey in chunk summary (#379)
We keep the summary of the index in memory, so that ZettaCache::lookup()
can find which chunk contains a given key.  We do this in a
`Vec<BlockBasedLogChunkSummaryEntry<IndexEntry>>`.  The IndexEntry is
composed of an IndexKey (9 bytes) and IndexValue (15 bytes).  However,
only the IndexKey is actually needed.

This commit changes the in-memory storage of the chunk summary to omit
the IndexValue, storing only the IndexKey.  This reduces the memory
usage from 32 to 17 bytes per chunk (53%).  For a 100TB cache, this
corresponds to a reduction of ~3GB RAM.

The interface to `SummarizedBlockBasedLog::lookup_by_key()` is also
simplified.

Note that the on-disk representation of the Summary is not changed; it
still (unnecessarily) stores the IndexValue.