DLPX-80708 Free beginning of IndexRun as we write new index (#380)
As we do a merge, the beginning of the old IndexRun is obsoleted by the
new IndexRun.  However, the old IndexRun is still stored in its
entirety.  This means that just before the merge completes, we have 2
whole copies of the Index, which is the largest piece of metadata in the
zettacache, typically around 1% of the cache size.

This commit changes the merge logic to free the beginning part of the
old IndexRun as we write the new one.

Reduce the amount of reserved slabs from 4% to 2%.