trace! memory logging: delayed formatting of DateTime
`trace!` statements are always logged to memory, and each one includes
the timestamp of when it was created.  The formatting (stingification)
of a DateTime is relatively expensive, and if we call `trace!`
frequently enough, this may impact performance.

This commit deferrs formatting of the DateTime until actually printing
it.  Until then it's stored as a `DateTime<Utc>`.

Bonus change: also delay formatting of the `Level`.