nice_p2size impl Display
nice_p2size, nice_number_count, and nice_number_time return String's,
which requires a memory allocation.

This commit changes these routines to return `impl Display`, a type
which can be later formatted with `{}`.  This avoids the allocation in
the common case where the routine is used as part of a `format!` (or
similar) macro.