check in object cache before zettacache (#406)
In `Pool::read_block()`, we check in the zettacache then in the object
cache, and then get the object from the object store.  This is an
inversion of the cache hierarchy, because the zettacache (disk) is
slower than the object cache (memory), which can lead to surprising poor
performance if a block is present in both the zettacache and object
cache.

This problem happens commonly if sibling block ingestion is enabled,
e.g. during sequential read with a cold zettacache.  On the read of the
first block of a given object, the object will be fetched from S3 and
added to the object cache, and all its blocks added to the zettacache.
Soon thereafter, the subsequent blocks in the object are read, which are
in both the zettacache and the object cache.

This commit changes the behavior such that `Pool::read_block()` first
checks if the block is already in the object cache or a GetObject is in
progress, and returns this data.  Then we check the zettacache and
finally get the object from S3 (via the object cache, but it's likely
not be there since we just checked a moment ago).

This has a big performance win for sibling block ingestion on certain
workloads, and is also better behavior for the default of single block
ingestion.

Bonus change: Now that we do write aggregation, it's rare to fill up the
ingestion buffer.  If sibling block ingestion is enabled, and we have
several concurrent read_block()'s in the same object (which is common),
the first one will actually perform the GetObject and see
`GetMethod::Loaded` and add all the object's blocks to the zettacache.
The concurrent read_block's will see `WaitedForInProgressLoad` and try
to insert the one block that they are accessing to the zettacache.
However, since all the blocks were just inserted, this is very likely to
be a no-op.  This commit changes it to not try to do this single-block
insertion when sibling block ingestion is enabled.

Bonus cleanup: Now that ingested buffers are normally copied to an
aligned, aggregate buffer in `Disk::aggregating_writer_thread()`,
there's no need to have the data in `AlignedBytes` beforehand.
Therefore, `ZettaCache::insert()` can take the `Bytes` directly rather
than using a callback to copy to an `AlignedBytes` only if we're
actually inserting it.  Instead we always provide a `Bytes`, which is
cheap to clone (it just bumps a refcount).