panic due to incorrect size in BlockBasedLogPhys::written_extents() (#413)
`BlockBasedLogPhys::iter_chunks()` uses `BlockBasedLogPhys::written_extents()`
to determine what extents contain valid (written and not trimmed)
chunks.  It then deserializes the chunks contained in these extents,
assuming that the extents are filled with valid chunks.

The problem is that if the log has only a single slab, and that slab is
trimmed, then `written_extents()` miscalculates the size of this extent,
returning a larger extent than was actually written.  `iter_chunks()`
then calls `BlockAccess::chunk_from_raw()` to deserialize a buffer that
was never written to, causing the agent to panic.

It's unlikely to get into this situation with the default tunable
values, because it's unlikely that we would take a checkpoint when the
merge is in the middle of the last slab of the index, and then the agent
crash after that checkpoint, requiring us to read the old index.

This commit corrects the calculation in `written_extents()`.

Note that `iter_chunks()` also has a check that tries to stop
deserializing chunks when we reach the last chunk (according to
`BlockBasedLogPhys::next_chunk`).  This check is not necessary because
we rely on `written_extents()` to only return valid extents, but it
would catch this error.  However, this check has an off-by-one error
introduced by the cleanup in #373.

This commit also removes the "last chunk" check in iter_chunks(), so we
rely fully on written_extents() to tell us where the end of the valid
chunks are.