use watch_once for Pool[Syncing]State::resuming (#403)
The `resuming` sender/receiver are used to indicate a one-time
transition from "resuming" to "not resuming" (or we may start in "not
resuming" if the agent is not restarting in the middle of a "txg sync"
operation).

This commit simplifies the logic for waiting for this transition to
complete, by taking advantage of the fact that
`watch_once::Receiver::recv()` will complete immediately if the sender
has been dropped.  Other kinds of channels also have this property; this
channel was chosen because its other uses are the most
simple/restrictive and therefore less prone to confusion.