make Progress::ingest() non-async (#387)
During index merging, `Progress::ingest()` is called very frequently
(once for each entry).  Since it's `async`, we need to instantiate its
Future for every call, which means initializing the future's storage
(async stack).  However, we only need to await() in the rare case that
we call `Progress::report()`, approximately once every million calls to
ingest().

This commit changes `Progress::ingest()` to be non-async.  It tells the
caller if we need to `report().await`.  This reduces CPU usage of the
merge_task() by 32% of one CPU.