aggregate BlockBasedLog writes in BlockBasedLog layer (#422)
Writes of metadata (BlockBasedLogs) are sequential on disk (within each
32MB slab).  We issue a write to each 8KB chunk, and these aggregated at
the BlockAccess layer into one pwrite() syscalls for each 1MB.  This has
a CPU cost which was thought to be small, but can become the bottleneck
for merge performance, especially when other changes are made to improve
merge performance.

This commit changes the write aggregation of BlockBasedLog chunks to
happen in the BlockBasedLog layer.  This saves one bcopy of the data,
and perhaps more importantly, reduces the number of messages that need
to be passed to the writer threads, and associated thread scheduling
overheads.

Compared to the pre-#344 code, this implementation is much cleaner
because it separates the concern of aggregation from the BlockBasedLog
syncing logic.

Bonus change: add the ability for `measure!` to optionally record timing
information.