ghost index size need not be limited (#429)
Previously, the amount of metadata was limited to the 10% of the disk
that was set aside when the zettacache was created.  Since this size was
fixed, we needed to ensure that the index size didn't grow too large.
However, now that metadata is allocated dynamically from the entire
cache, this is no longer the case.

This commit removes the limit on the GHOST_CACHE_SIZE_PCT tunable.