DLPX-81169 zcache hits overall hit count/percent is inaccurate (#430)
`zcache hits` prints the number of cache hits, and the percent of
lookups that are hits.  However, the value printed is not accurate - it
includes “ghost” hits, which do not actually return data to the caller,
they just found an entry in the index for a recently evicted block.
Notably, in the graph there are never any `*`'s after the `|`,
incorrectly indicating that the cache size is double what it actually
is, and that that increasing the cache size will not increase the hit
rate.

This commit corrects the problem by having the agent tell the `zcache`
command how many actual (real, live) hits there were.

Additional improvements:

The `--quantiles` / `-q` option now specifies the total number of
quantiles to display, instead of the number of quantiles of "live" data.

Display start and end time in the local timezone

Display elapsed time between start and end

Set histogram width and length (quantiles) based on terminal size (or
default to 80x24 if output is not a tty)

Display hit percent with one decimal place

Code cleanup, e.g. renaming variables