simplify `zcache` error handling (#436)
The agent's result that's returned to `zcache` can include an error
enum.  However, `zcache` assumes that any error is due to the zettacache
not being present.

This commit changes the agent to return a simple error string, in
addition to a potential error enum.  If the `zcache` command receives an
error response from the agent, it will print out the simple error string
and exit.

When making a request on a non-existent zettacache, the wording of the
error message is slightly changed.  However, note that this error case
will be removed entirely in the future, treating a zettacache with no
devices as an empty zettacache (with no disks, no hits, etc).

Bonus fix: if the `zcache` process doesn't have permission to connect to
the agent's socket, fail immediately rather than retrying for 15
seconds.  This is typical when running without `sudo` and the subcommand
requires root privileges, e.g. `zcache hits --clear`, `zcache add`, etc.