zettacache index merge performance improvements (#424)
In `MergeState::merge_task`, we need to mergesort the old index with the
PendingChanges to produce the new index.  This involves iterating through both
the old index and the PendingChanges, taking whichever entry is lower.  We do
this with two nested loops, the outer loop iterating over the PendingChanges,
and the inner loop advancing through the old index.  So for each entry in the
old index, we compare it with the next PendingChange to determine which should
be handled next.

There are typically many more entries in the old index (which is on disk) than
in the PendingChanges (which is in memory).  This commit improves performance
by restructuring the code so that the outer loop iterates over the old index,
and the inner loop iterates over the PendingChanges.  We can determine the
contiguous run of entries in the old index which are all before the next
pending change in O(log(n)), and in general this contiguous run is likely to be
large.  This way, the hottest path minimizes the amount of branches and memory
accesses, since we don't have to look at the next PendingChange for every entry
in the old index.  The resulting code is also arguable easier to follow.

Additionally, some micro optimizations are made to reduce CPU usage of the
hottest functions (e.g. `Progress::obsolete()` and `Progress::ingest()`):

In `AtimeHistogramPhys`, introduce `insert_unchecked()` and `with_capacity()`
to reduce branches compared to `insert()`.

`IndexRun::append()`: have the caller compute the `AtimeHistogramPhys` so that
we don't need to iterate the `Vec<IndexEntry>` here.  The caller is already
iterating each entry.

Use `Atime.0 >= Atime.0` so that the primitive u32 comparison is used, which
the compiler can better optimize compared to calling
`<Atime as PartialOrd>::ge()`.

Performance of the merge increases from 21 -> 27 million entries per
second (+28%).  The index is written at ~585MiB/s, which is close to the
hardware maximum of this machine which is around 610MiB/s.