live zettacache creation (#442)
Currently, disks can be added to a zettacache while the agent is
running, with `zcache add DISK`.  However, initial zettacache creation
can be done only when the agent is started, with
`zfs_object_agent -c DISK`.

This commit adds the ability to create a new zettacache with
`zcache add DISK`, "adding" the first disk to an empty zettacache.

The ability to create, add, and expand disks with `zfs_object_agent -c`
remains.  These will be removed in the future, after implementing live
expansion and changing the product to use the "live" capabilities.

In terms of the implementation, the agent now always has exactly one
instance of `struct ZettaCache`, even if there are no disks in the
zettacache.  If there are disks in the zettacache, then there will be
one instance each of `struct Inner` (formerly `ZettaCache`) and
`struct Locked` (formerly `ZettaCacheState`).