zcache expand (#447)
Currently, zcache devices are automatically expanded when the agent
starts.

This commit adds that ability to expand zcache devices while the agent
is running, with the new `zcache expand DISK` subcommand.

Note that we still auto-expand on open.  In the future we'll make the
appstack use `zcache expand`, and then we can remove the auto-expand
code.