zoa: determine response type automatically (#446)
Since the `response_type` is the same as the `request_type`, the server
infrastructure can fill it in automatically, so each request handler
doesn't need to match it up manually.

The "get pools" response contained the pools directly in the nvlist, and
asserted that there are no other nvpairs.  There was no `response_type`
in this response message.  This commit changes it so that "get pools"
includes a `response_type`, and the pools are under a different `pools`
nvpair.  Additionally, the code in `PublicConnectionState::get_pools()`
is reorganized to be more clear.

Additionally, the code in `PublicConnectionState::get_destroying_pools()`
is simplified, using the server helper method to convert to an nvlist.