increase zfs send prefetching (#456)
When `zfs send` has to retrieve the data from the object store (as
opposed to zettacache), we can benefit from a larger prefetch queue
depth, similar to what we see for predictive prefetching.  This is
because we need a high GetObject queue depth (dozens of requests,
covering dozends of MB's of data) to get good throughput from S3.

This commit increases the zfs send prefetching to 64MB (from 16MB), and
clarifies the comment around it (there's really no downside to
increasing it aside from RAM usage, since this is not using ARC
prefetching any more).