Arc not needed for indices and Locked (#481)
We don't really need to have separate refcounts for the indices and the
Locked; instead we can use the Inner refcount where necessary.

We also don't need to have separate RwLock's for the two indices;
instead we can use a single lock which contains both old and new
indices.

One of the few places that we cloned these Arc's separately is for
`MergeState::merge_task()`, which takes
`old_index_lock: Arc<tokio::sync::RwLock<IndexRun>>`.  This commit
changes it to take `old_index_phys: IndexRunPhys`.  A few accessor
methods need to be added to the IndexRunPhys to facilitate this.