Object Store tests hang in github CI
The zfs-object-agent service is used by ZTS, but that service depends
on the zfs-mount service, which isn't enabled by the package install
done by CI. This causes the agent to fail to start, and the rest of the
test run hangs. By enabling and starting zfs-mount prior to launching
the tests, the agent can successfully start.